Fix Collision errors
knockback values need tuning
