+++ Assets/WumpusCollisionScript.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 53bbf14514aa6354382f917ecd8ad495
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
